-added exponential_moving_average logic -narrowed window type

+++ include/exponential_moving_average.h
@@ -11,6 +11,6 @@ public:
     ExponentialMovingAverage() = default;
 
     [[nodiscard]]
-    std::map<std::string, double> calculate(const std::vector<OhlcData> &ohlc_data, size_t window) const override;
+    std::map<std::string, double> calculate(const std::vector<OhlcData> &ohlc_data, int window) const override;
 };
 #endif //EXPONENTIAL_MOVING_AVERAGE_H
